# frontend/src/api/openapi-snapshot.json
{
  "components": {
    "schemas": {
...
      "AnomalyRead": {
        "additionalProperties": false,
        "description": "Une anomalie de conformité.\n\n`focus` est un point du **plan** en centimètres, dans le repère de `Room.polygon` : c'est ce\nqui permet au panneau de recentrer l'éditeur 2D sans conversion. `null` quand l'anomalie ne\ndésigne aucun endroit précis — une pièce sans ouverture n'a pas de coordonnée.",
        "properties": {
          "element_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Element Ids",
            "type": "array"
          },
          "face_labels": {
            "items": {
              "type": "string"
            },
            "title": "Face Labels",
            "type": "array"
          },
          "focus": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Focus"
          },
          "measured_cm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Measured Cm"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "room_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Room Id"
          },
          "room_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Room Name"
          },
          "rule_id": {
            "title": "Rule Id",
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "threshold_cm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threshold Cm"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "rule_id",
          "title",
          "severity",
          "message"
        ],
        "title": "AnomalyRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InspectionRead": {
        "additionalProperties": false,
        "description": "Le rapport complet.\n\n`thresholds` republie les seuils appliqués : un rapport qui annonce « passage insuffisant »\nsans dire par rapport à quoi n'est pas vérifiable, et le mode accessible change la réponse.",
        "properties": {
          "anomalies": {
            "items": {
              "$ref": "#/components/schemas/AnomalyRead"
            },
            "title": "Anomalies",
            "type": "array"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Counts",
            "type": "object"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "rooms": {
            "items": {
              "$ref": "#/components/schemas/RoomInspectionRead"
            },
            "title": "Rooms",
            "type": "array"
          },
          "thresholds": {
            "additionalProperties": true,
            "title": "Thresholds",
            "type": "object"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "thresholds",
          "rooms",
          "anomalies",
          "counts",
          "warnings"
        ],
        "title": "InspectionRead",
        "type": "object"
      },
...
      "LayingPlanRead": {
        "additionalProperties": false,
        "description": "Enveloppe du calepinage optimisé.\n\nLe détail par face reste un dictionnaire libre, pour la même raison que le métré : sa forme\ndépend du revêtement, une face peinte n'a pas de calepinage du tout, et un `Union` de variantes\nalourdirait le schéma sans rien apporter au client. Ce qui est typé, c'est ce qui se lit :\n`cuts_saved` est le chiffre que l'artisan compare.",
        "properties": {
          "cuts_saved": {
            "title": "Cuts Saved",
            "type": "integer"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "rooms": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Rooms",
            "type": "array"
          }
        },
        "required": [
          "rooms",
          "cuts_saved"
        ],
        "title": "LayingPlanRead",
        "type": "object"
      },
      "LayoutItemRead": {
        "additionalProperties": false,
        "description": "Un meuble proposé, prêt à être créé tel quel.\n\n`pos_x_cm` / `pos_y_cm` sont le **centre** de l'emprise dans le repère du plan, la convention\ndu mobilier libre (`docs/spec-complete.md` §10, amendement A4). Le client n'a donc aucune\nconversion à faire pour transformer une proposition en éléments réels.",
        "properties": {
          "against_face_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Against Face Label"
          },
          "clearance_cm": {
            "title": "Clearance Cm",
            "type": "number"
          },
          "depth_cm": {
            "title": "Depth Cm",
            "type": "number"
          },
          "height_cm": {
            "title": "Height Cm",
            "type": "number"
          },
          "pos_x_cm": {
            "title": "Pos X Cm",
            "type": "number"
          },
          "pos_y_cm": {
            "title": "Pos Y Cm",
            "type": "number"
          },
          "rotation_deg": {
            "title": "Rotation Deg",
            "type": "number"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "width_cm": {
            "title": "Width Cm",
            "type": "number"
          }
        },
        "required": [
          "slug",
          "width_cm",
          "depth_cm",
          "height_cm",
          "pos_x_cm",
          "pos_y_cm",
          "rotation_deg",
          "clearance_cm"
        ],
        "title": "LayoutItemRead",
        "type": "object"
      },
      "LayoutProposalRead": {
        "additionalProperties": false,
        "description": "Une implantation valide, avec le détail de sa note.\n\n`breakdown` n'est pas de la décoration : sans lui, le classement est une boîte noire et\nl'utilisateur n'a aucun moyen de comprendre pourquoi la deuxième proposition perd.",
        "properties": {
          "breakdown": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Breakdown",
            "type": "object"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/LayoutItemRead"
            },
            "title": "Items",
            "type": "array"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "score": {
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "rank",
          "score",
          "breakdown",
          "items"
        ],
        "title": "LayoutProposalRead",
        "type": "object"
      },
      "LayoutProposalsRead": {
        "additionalProperties": false,
        "properties": {
          "program": {
            "title": "Program",
            "type": "string"
          },
          "proposals": {
            "items": {
              "$ref": "#/components/schemas/LayoutProposalRead"
            },
            "title": "Proposals",
            "type": "array"
          },
          "room_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Room Id"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          },
          "weights": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Weights",
            "type": "object"
          }
        },
        "required": [
          "program",
          "weights",
          "proposals",
          "warnings"
        ],
        "title": "LayoutProposalsRead",
        "type": "object"
      },
      "LayoutRequest": {
        "additionalProperties": false,
        "description": "Ce qu'un client peut demander. Rien d'autre : `extra=\"forbid\"`.\n\nAucun seuil n'entre par le corps de la requête. Les rendre pilotables par le client\ntransformerait un contrôle métier en paramètre d'affichage, et il suffirait de demander un\nseuil de 10 cm pour qu'un plan invivable devienne conforme.",
        "properties": {
          "accessible": {
            "default": false,
            "title": "Accessible",
            "type": "boolean"
          },
          "count": {
            "default": 3,
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Count",
            "type": "integer"
          },
          "program": {
            "anyOf": [
              {
                "pattern": "^[a-z_]{1,40}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Program"
          }
        },
        "title": "LayoutRequest",
        "type": "object"
      },
...
      "RoomInspectionRead": {
        "additionalProperties": false,
        "properties": {
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Counts",
            "type": "object"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "room_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Room Id"
          }
        },
        "required": [
          "counts"
        ],
        "title": "RoomInspectionRead",
        "type": "object"
      },
...
      "Severity": {
        "description": "Trois niveaux, et un seul mérite d'arrêter un chantier.",
        "enum": [
          "bloquant",
          "avertissement",
          "conseil"
        ],
        "title": "Severity",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Éditeur de plan de rénovation 2D → 3D",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/projects/{project_id}/inspection": {
      "get": {
        "description": "Toutes les anomalies de conformité du projet, triées par sévérité puis par pièce.\n\nChacune porte un identifiant de règle stable, une mesure, le seuil qu'elle enfreint et les\nentités concernées : c'est ce qui permet au panneau d'inspection de recentrer le plan sur le\nproblème plutôt que de laisser l'utilisateur le chercher.\n\n`warnings` recense ce qui **n'a pas pu** être contrôlé — une pièce dont les murs ne se\nreferment pas n'a pas de contour, donc pas de circulation mesurable. Un rapport vide\naccompagné d'un avertissement ne veut pas dire « conforme ».",
        "operationId": "read_inspection_api_projects__project_id__inspection_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "integer"
            }
          },
          {
            "description": "Applique les seuils d'un logement accessible (couloir de 120 cm) en plus des seuils d'usage courant.",
            "in": "query",
            "name": "accessible",
            "required": false,
            "schema": {
              "default": false,
              "description": "Applique les seuils d'un logement accessible (couloir de 120 cm) en plus des seuils d'usage courant.",
              "title": "Accessible",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Inspection",
        "tags": [
          "intelligence"
        ]
      }
    },
    "/api/projects/{project_id}/laying-plan": {
      "get": {
        "description": "Calepinage optimisé : sens de pose, position de la première rangée, et plinthes.\n\nLe métré (`GET /takeoff`) donne déjà entières, coupes et chutes pour la pose de référence —\ntrame calée sur le coin, unité dans le sens déclaré. Cette route dit **comment poser** : quel\nsens minimise les coupes, où démarrer la trame pour ne jamais laisser moins d'un tiers d'unité\nen rive, et combien de barres de plinthe commander en réemployant les chutes.\n\nElle n'entre pas dans le devis et ne le contredit pas : `cuts_saved` mesure l'écart avec la\npose de référence du métré, il ne réécrit pas la quantité à commander, qui reste établie sur la\nsurface et le taux de chute.",
        "operationId": "read_laying_plan_api_projects__project_id__laying_plan_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LayingPlanRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Laying Plan",
        "tags": [
          "intelligence"
        ]
      }
    },
...
    "/api/rooms/{room_id}/layouts": {
      "post": {
        "description": "Deux ou trois implantations de mobilier **valides** et classées, pour la pièce demandée.\n\nUn `POST` alors que rien n'est écrit : le verbe est celui du corps de requête, pas celui d'un\neffet de bord. Aucune ligne n'est créée — le client choisit une proposition et crée lui-même\nles éléments, parce qu'un moteur qui poserait d'autorité quinze meubles dans le plan d'un\nartisan serait un moteur qu'on désactive au premier essai.\n\nLe rôle exigé est `editor` et non `viewer`, pour deux raisons : la proposition n'a de sens que\npour quelqu'un qui peut ensuite modifier le plan, et c'est le calcul le plus coûteux de l'API —\nl'ouvrir en lecture seule en ferait le levier de charge le moins cher du produit.",
        "operationId": "propose_room_layouts_api_rooms__room_id__layouts_post",
        "parameters": [
          {
            "in": "path",
            "name": "room_id",
            "required": true,
            "schema": {
              "title": "Room Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LayoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LayoutProposalsRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Propose Room Layouts",
        "tags": [
          "intelligence"
        ]
      }
    },
...
  }
}